refactor(pipeline/network): extract FrozenReplay; NetworkDiscovery becomes a 24-LoC shim
Move the final two functions out of NetworkDiscovery.ts into a
dedicated FrozenReplay/ sub-module:

  • createFrozenNetwork — synthesises a frozen INetworkDiscovery
    from a static endpoint snapshot. SCRAPE.ACTION uses it to replay
    captured pools without holding a Playwright Page open.
  • buildFrozenHeaders — produces the header bundle bound to the
    pre-cached auth token, mirroring the live engine's discovery
    rules so live and frozen paths agree on origin / siteId / referer
    propagation.

NetworkDiscovery.ts shrinks to a 24-LoC compatibility shim that
re-exports the public symbols (createNetworkDiscovery,
createFrozenNetwork, distillHeaders, parseTextOrNull,
shouldRecordResponse, isUnsupportedUrl, parseResponse and the typed
signals + endpoint/discovery types) from their new homes. All 50
existing import sites continue to work without changes.

Phase 4 split: commit 7+8/9 — folded together because the
FrozenReplay extraction already reduces NetworkDiscovery.ts to a pure
re-export facade, making the planned separate shim commit redundant.
NetworkDiscovery.ts: 179 → 24 LoC (baseline 1812).
